Added a new redis based cache-system to propagate generated shared schama objects to all pods on openshift.

+++ libs/store/api-core/src/lib/services/api-util-service.ts
@@ -1,9 +1,10 @@
-import { Observable, throwError } from 'rxjs';
+import { Observable, from, of, throwError } from 'rxjs';
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
-import { catchError, map } from 'rxjs/operators';
+import { catchError, map, switchMap } from 'rxjs/operators';
 import { OpenAPIV3 } from 'openapi-types';
 import { StoreConfigService } from '@luomus/store/config';
 import { FileService, UtilityService } from '@luomus/store/shared';
+import { SchemaCacheService } from '@luomus/store/schema-cache';
 
 @Injectable()
 export class ApiUtilService {
@@ -21,33 +22,63 @@ export class ApiUtilService {
 
   constructor(
     private fileService: FileService,
-    private configService: StoreConfigService
+    private configService: StoreConfigService,
+    private schemaCacheService: SchemaCacheService
   ) {}
 
   getJsonLDContext(type: string) {
-    return this.getFile(
-      this.fileService.getFilename(
-        UtilityService.removeSuffix(type, '.json'),
-        this.configService.get('JSON_LD_CONTEXT_PATH')
-      ),
-      'JsonLD Context not found'
+    type = UtilityService.removeSuffix(type, '.json');
+
+    return from(this.schemaCacheService.getCachedJsonLd(type)).pipe(
+      switchMap(data => {
+        if (data) {
+          return of(data);
+        } else {
+          return this.getFile(
+            this.fileService.getFilename(
+              UtilityService.removeSuffix(type, '.json'),
+              this.configService.get('JSON_LD_CONTEXT_PATH')
+            ),
+            'JsonLD Context not found'
+          );
+        }
+      })
     );
   }
 
   getJsonSchema(type: string): Observable<OpenAPIV3.Document> {
-    return this.getFile(
-      this.fileService.getFilename(UtilityService.removeSuffix(type, '.json')),
-      'Schema not found'
+    type = UtilityService.removeSuffix(type, '.json');
+
+    return from(this.schemaCacheService.getCachedJsonSchema(type)).pipe(
+      switchMap(data => {
+        if (data) {
+          return of(data);
+        } else {
+          return this.getFile(
+            this.fileService.getFilename(UtilityService.removeSuffix(type, '.json')),
+            'Schema not found'
+          );
+        }
+      })
     );
   }
 
   getEsMapping(type: string) {
-    return this.getFile(
-      this.fileService.getFilename(
-        UtilityService.removeSuffix(type, '.json'),
-        this.configService.get('ES_INDEX_PATH')
-      ),
-      'Elasticsearch mapping not found'
+    type = UtilityService.removeSuffix(type, '.json');
+    return from(this.schemaCacheService.getCachedEsIndex(type)).pipe(
+      switchMap(data => {
+        if (data) {
+          return of(data);
+        } else {
+          return this.getFile(
+            this.fileService.getFilename(
+              type,
+              this.configService.get('ES_INDEX_PATH')
+            ),
+            'Elasticsearch mapping not found'
+          )
+        }
+      })
     ).pipe(map((index) => index?.mappings));
   }
